Dodan User list za admina
- Podijeljen /admin page na unit i users.
- U users tabu možeš vidjeti sve korisnike u bazi i njihovu najvišu ulogu.
- Dodana mogućnost brisanja korisnika. (bilo problema s tokenima jer se nisu spremali tokom logina sad se spremaju i brisu tokom logouta)

Napravljeno: 8.11.
Vrijeme potrošeno: ~1 sat

+++ backend/src/main/java/springboot/backend/controller/PersonController.java
@@ -76,9 +76,20 @@ public class PersonController {
         return repo.findById(id);
     }
 
-
     @DeleteMapping("/deletePerson/{id}")
-    public void deletePerson(@PathVariable Long id) {
+    public ResponseEntity<?> deletePerson(@PathVariable Long id, @AuthenticationPrincipal Jwt jwt) {
+        String email = jwt.getClaimAsString("email");
+        Optional<Person> current = repo.findByEmail(email);
+
+        if (current.isEmpty()) return ResponseEntity.status(403).body("Unauthorized");
+
+        Person currentUser = current.get();
+        if (currentUser.getId().equals(id)) {
+            return ResponseEntity.status(403).body("You cannot delete yourself!");
+        }
+
         repo.deleteById(id);
+        return ResponseEntity.ok("User deleted");
     }
+
 }
